requirements.txt
Alfred-Workflow==1.36
pyotp==2.2.7
